+++ website/src/pages/index.js
@@ -16,6 +16,8 @@ import LanguageInfoWidget from '../components/Homepage/LanguageInfoWidget';
 import RiverSection from '../components/Homepage/RiverSection';
 import ThreeCardsWithIcon from '../components/Homepage/ThreeCardsWithIcon';
 
+import HomePageExample from '!!raw-loader!./home_page_example.py';
+
 function GetStartedSection() {
     return (
         <section className={styles.languageGetStartedSection}>
@@ -28,57 +30,12 @@ function GetStartedSection() {
     );
 }
 
-const example = `import asyncio
-
-from crawlee.crawlers import PlaywrightCrawler, PlaywrightCrawlingContext
-
-
-async def main() -> None:
-    crawler = PlaywrightCrawler(
-        max_requests_per_crawl=5,  # Limit the crawl to 5 requests at most.
-        headless=False,  # Show the browser window.
-        browser_type='firefox',  # Use the Firefox browser.
-    )
-
-    # Define the default request handler, which will be called for every request.
-    @crawler.router.default_handler
-    async def request_handler(context: PlaywrightCrawlingContext) -> None:
-        context.log.info(f'Processing {context.request.url} ...')
-
-        # Extract and enqueue all links found on the page.
-        await context.enqueue_links()
-
-        # Extract data from the page using Playwright API.
-        data = {
-            'url': context.request.url,
-            'title': await context.page.title(),
-            'content': (await context.page.content())[:100],
-        }
-
-        # Push the extracted data to the default dataset.
-        await context.push_data(data)
-
-    # Run the crawler with the initial list of URLs.
-    await crawler.run(['https://crawlee.dev'])
-
-    # Export the entire dataset to a JSON file.
-    await crawler.export_data('results.json')
-
-    # Or work with the data directly.
-    data = await crawler.get_data()
-    crawler.log.info(f'Extracted data: {data.items}')
-
-
-if __name__ == '__main__':
-    asyncio.run(main())
-`;
-
 function CodeExampleSection() {
     return (
         <section className={styles.codeExampleSection}>
             <div className={styles.decorativeRow} />
             <div className={styles.codeBlockContainer}>
-                <CodeBlock language="python">{example}</CodeBlock>
+                <CodeBlock language="python">{HomePageExample}</CodeBlock>
             </div>
             <div className={styles.dashedSeparator} />
             <div className={styles.decorativeRow} />
